refactor: 연산 대상이 되는 문자열과 구분자 추출 로직을 InputProcessor 클래스로 분리
